Renamed auto file
