+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/libraries/Motor.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.auto;
+package org.firstinspires.ftc.teamcode.libraries;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
